+++ templates/Module/ignoreds.php
@@ -1,0 +1,4 @@
+<?php
+
+foreach ($this->ignoredModules as $moduleName)
+    echo $moduleName . PHP_EOL;